Features File Access and redis for authentication
- Implemented the file Access feature
- Imlmented redis for verification
- Commented the entire code

+++ src/controllers/fileController.js
@@ -1,13 +1,22 @@
 const fs = require('fs');
 const path = require('path');
-const { FileService, DirectoryService, FileMapService, FileVersionService } = require('../services');
+const {
+  UserService,
+  FileService,
+  DirectoryService,
+  FileMapService,
+  FileAccessService,
+  FileVersionService,
+} = require('../services');
 const { Response, Message, fileHashGenerator, Logger } = require('../utilities');
 
 class FileController {
+  // Validate and process file data before insertion
   static async _validateAndProcessFile(fileData, creatorId, directoryId = null) {
     try {
       const { file, ...metaData } = fileData;
       const checkParams = { directoryId, creatorId };
+
       // Generate a hash key from the file stream
       const fileHash = await fileHashGenerator.generateHashKey(file.path);
 
@@ -18,6 +27,7 @@ class FileController {
         await FileMapService.isUnique(checkParams);
       }
 
+      // Prepare file parameters for insertion
       const fileParams = {
         fileName: file.filename,
         filePath: file.path,
@@ -31,6 +41,7 @@ class FileController {
 
       return fileParams;
     } catch (err) {
+      // If any error occurs, unlink the file to prevent storage leaks
       await fs.promises.unlink(fileData.file.path);
       throw err;
     }
@@ -41,6 +52,7 @@ class FileController {
       const fileData = req.body;
       const fileParams = await FileController._validateAndProcessFile(fileData, req.user.userId, null);
 
+      // Insert file into database
       const fileRes = await FileMapService.insert(fileParams);
       Response.success(res, fileRes, Message.FILE_UPLOADED);
     } catch (err) {
@@ -53,10 +65,12 @@ class FileController {
       const fileData = req.body;
       const { directoryId } = req.params;
 
+      // Check if the directory exists and user has permission to upload
       await DirectoryService.checkDirectoryExistence({ directoryId, creatorId: req.user.userId });
 
       const fileParams = await FileController._validateAndProcessFile(fileData, req.user.userId, directoryId);
 
+      // Insert file into database
       const fileRes = await FileMapService.insert(fileParams);
       Response.success(res, fileRes, Message.FILE_UPLOADED);
     } catch (err) {
@@ -66,13 +80,17 @@ class FileController {
 
   static async downloadFile(req, res, next) {
     try {
-      const { fileId } = req.params;
+      const { id } = req.params;
 
-      const file = await FileMapService.getById(fileId);
+      // Fetch file details
+      const file = await FileMapService.getById(id);
 
-      if (file.creatorId !== req.user.userId) {
-        // check file permission
+      // Check user permission to access the file
+      if (file.accessType === 'private' && file.creatorId !== req.user.userId) {
         throw Response.createError(Message.ACCESS_DENIED);
+      } else if (file.accessType === 'partial') {
+        // Check file permission file access
+        await FileAccessService.checkUserFileAccess({ id, userId: req.user.userId });
       }
 
       // Check if the file exists
@@ -80,15 +98,16 @@ class FileController {
 
       // Create a readable stream to read the file
       const fileStream = fs.createReadStream(file.filePath);
-      // Handle errors during streaming
+      // Handle streaming errors
       fileStream.on('error', (err) => {
         Logger.error('Error streaming file:', err);
         throw Response.createError(Message.GENERIC_ERROR, err);
       });
 
-      // Set the appropriate content type and attachment header for the file
+      // Set response headers for file download
       res.setHeader('Content-Type', 'application/octet-stream');
       res.setHeader('Content-Disposition', `attachment; filename="${file.fileName}.${file.fileType}"`);
+
       // Pipe the file stream to the response stream
       fileStream.pipe(res);
     } catch (err) {
@@ -100,58 +119,64 @@ class FileController {
     try {
       const params = { ...req.body, ...req.params };
 
-      // check file
-      const file = await FileMapService.getById(req.params.fileId);
-
+      // Check file existence and user permission
+      const file = await FileMapService.getById(params.id);
       if (file.creatorId !== req.user.userId) {
         throw Response.createError(Message.ACCESS_DENIED);
       }
 
+      // Update file metadata
       const srvRes = await FileMapService.updateMetaData(params);
-
       Response.success(res, srvRes, 'MetaData updated');
     } catch (err) {
       next(err);
     }
   }
 
+  // Delete file
   static async delete(req, res, next) {
     try {
-      const params = { ...req.params };
-      params.creatorId = req.user.userId;
+      const { id } = { ...req.params };
 
-      await FileMapService.getById(req.params.fileId);
+      const params = {
+        id,
+        creatorId: req.user.userId,
+      };
+
+      // Check file existence and user permission
+      const file = await FileMapService.getById(id);
+      if (file.creatorId !== params.creatorId) {
+        throw Response.createError(Message.ACCESS_DENIED);
+      }
 
       const srvRes = await FileMapService.delete(params);
-
       Response.success(res, srvRes, 'File deleted');
     } catch (err) {
       next(err);
     }
   }
 
-  // versions
+  // File version operations
+  // Get all versions of a file
   static async getFileVersions(req, res, next) {
     try {
-      const { fileId } = req.params;
+      const { id } = req.params;
 
-      const fileRes = await FileVersionService.getAll(fileId);
-
+      const fileRes = await FileVersionService.getAll(id);
       Response.success(res, fileRes, fileRes ? 'Version file(s) found' : 'No  versions found for this file');
     } catch (err) {
       next(err);
     }
   }
 
+  // Upload a new version of a file
   static async uploadVersion(req, res, next) {
     try {
       const fileData = req.body;
-      const { fileId } = req.params;
+      const { id } = req.params;
 
-      // Check if the file exists
-      const existingFile = await FileMapService.getById(fileId, true);
-
-      // Check if the user has permission to upload a new version of the file
+      // Check file existence and user permission
+      const existingFile = await FileMapService.getById(id, true);
       if (existingFile.creatorId !== req.user.userId) {
         throw Response.createError(Message.ACCESS_DENIED);
       }
@@ -162,16 +187,79 @@ class FileController {
         existingFile.directoryId,
       );
 
-      // file hash check
+      // Check if the uploaded version is a duplicate
       if (fileParams.fileHash === existingFile.fileHash) {
         throw Response.createError(Message.DUPLICATE_FILE);
       } else {
-        await FileVersionService.checkHashKeyExists(fileId, fileParams.fileHash);
+        await FileVersionService.checkHashKeyExists(id, fileParams.fileHash);
       }
 
+      // Insert new version
       const fileRes = await FileVersionService.insert(fileParams, existingFile);
-
       Response.success(res, fileRes, Message.FILE_VERSION_UPLOADED);
+    } catch (err) {
+      next(err);
+    }
+  }
+
+  // Restore a specific version of a file
+  static async restoreVersion(req, res, next) {
+    try {
+      const { id, versionId } = req.params;
+
+      // Check file existence and user permission
+      const existingFile = await FileMapService.getById(id, true);
+      if (existingFile.creatorId !== req.user.userId) {
+        throw Response.createError(Message.ACCESS_DENIED);
+      }
+
+      // Check if the version exists and restore it
+      const version = await FileVersionService.checkFileVersion({ id, versionId });
+      await FileVersionService.restore(version);
+
+      Response.success(res, {}, Message.VERSION_RESTORED);
+    } catch (err) {
+      next(err);
+    }
+  }
+
+  // File Access operations
+  static async setAccess(req, res, next) {
+    try {
+      const { id } = req.params;
+      const { accessType, allowedUserIds = [] } = req.body;
+
+      // Check file existence and user permission
+      const existingFile = await FileMapService.getById(id);
+      if (existingFile.creatorId !== req.user.userId) {
+        throw Response.createError(Message.ACCESS_DENIED);
+      }
+
+      if (allowedUserIds) {
+        await UserService.validateUserIds(allowedUserIds);
+      }
+
+      await FileAccessService.setAccess({ accessType, allowedUserIds, file: existingFile });
+      Response.success(res, {}, Message.FILE_ACCESS_UPDATED);
+    } catch (err) {
+      next(err);
+    }
+  }
+
+  static async removeUserAccess(req, res, next) {
+    try {
+      const { id, userId } = req.params;
+
+      // Check file existence and user permission
+      const existingFile = await FileMapService.getById(id);
+      if (existingFile.creatorId !== req.user.userId) {
+        throw Response.createError(Message.ACCESS_DENIED);
+      } else if (existingFile.accessType !== 'partial') {
+        throw Response.createError(Message.ACCESS_DENIED);
+      }
+
+      await FileAccessService.removeUserAccess({ id, userId });
+      Response.success(res, {}, Message.FILE_ACCESS_UPDATED);
     } catch (err) {
       next(err);
     }
